+++ CHANGELOG.md
@@ -4,12 +4,19 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
-## [x.y.z] - YYYY-MM-DD
+### Added
+- Duplicate object checks in the Resolution sensemaker are now configurable 
 
-> *Galadriel-INC-2?*
+### Changed
+- None
+
+## [0.2.0] - 2025-03-20
+
+> *Galadriel-INC-2*
 
 ### Added
-- None 
+- Mil Symbol now updates the node's symbolIdCode with the 2525-C SDIC
+- Mil Symbol enrichment can now look at controlling parent node's affiliation
 
 ### Changed
 - None
@@ -27,5 +34,6 @@ All notable changes to this project will be documented in this file.
 
 
 
-[Unreleased]: https://tex.gerbil-cloud.ts.net:3000/oms/oms-sensemaking/compare/v0.1.0...main
+[Unreleased]: https://tex.gerbil-cloud.ts.net:3000/oms/oms-sensemaking/compare/v0.2.0...main
+[0.2.0]: https://tex.gerbil-cloud.ts.net:3000/oms/oms-sensemaking/releases/tag/v0.2.0
 [0.1.0]: https://tex.gerbil-cloud.ts.net:3000/oms/oms-sensemaking/releases/tag/v0.1.0